Add updates from Demo branch, move to new generic divertor designer location, make minor changes ready for use on other devices
