+++ eudemo/eudemo_tests/ivc/test_divertor_silhouette.py
@@ -40,7 +40,9 @@ class TestDivertorSilhouetteDesigner:
         "div_Ltarg_ob": {"value": 0.5, "unit": "m"},
         "div_targ_angle_ib": {"value": 42, "unit": "degrees"},
         "div_targ_angle_ob": {"value": -25, "unit": "degrees"},
-        "div_open": {"value": False, "unit": ""},
+        "div_targ_type_ib": {"value": "verticle", "unit": ""},
+        "div_targ_type_ob": {"value": "verticle", "unit": ""},
+        "div_baffle_type": {"value": "circle_baffle", "unit": ""},
     }
 
     @classmethod
@@ -82,7 +84,6 @@ class TestDivertorSilhouetteDesigner:
         params["div_targ_angle_ib"]["value"] = div_targ_angle
 
         designer = DivertorSilhouetteDesigner(params, self.eq, self.wall)
-
         divertor = designer.execute()
 
         for target in [divertor[1], divertor[3]]:
@@ -100,8 +101,8 @@ class TestDivertorSilhouetteDesigner:
 
         _, inner_target, dome, outer_target, _ = designer.execute()
 
-        assert signed_distance(dome, inner_target) == 0
-        assert signed_distance(dome, outer_target) == 0
+        assert signed_distance(dome, inner_target) == pytest.approx(0)
+        assert signed_distance(dome, outer_target) == pytest.approx(0)
 
     def test_dome_does_not_intersect_separatrix(self):
         designer = DivertorSilhouetteDesigner(self.params, self.eq, self.wall)
@@ -121,7 +122,9 @@ class TestDivertorSilhouetteDesigner:
         assert len(turning_points) == 1
         assert dome_coords[2, turning_points[0]] < x_points[0].z
 
-    def test_baffle_start_and_end_points_and_target_intersects(self):
+    @pytest.mark.parametrize("div_baffle_type", ["circle_baffle", "fluxline_baffle"])
+    def test_baffle_start_and_end_points_and_target_intersects(self, div_baffle_type):
+        self.params["div_baffle_type"]["value"] = div_baffle_type
         designer = DivertorSilhouetteDesigner(self.params, self.eq, self.wall)
 
         inner_baffle, inner_target, _, outer_target, outer_baffle = designer.execute()
