+++ eudemo/eudemo_tests/template.json
@@ -821,8 +821,14 @@
     "unit": "m",
     "source": "Input"
   },
-  "div_Ltarg": {
-    "name": "Divertor target length",
+  "div_Ltarg_ib": {
+    "name": "Divertor target inboard length",
+    "value": 0.5,
+    "unit": "m",
+    "source": "Input"
+  },
+  "div_Ltarg_ob": {
+    "name": "Divertor target outboard length",
     "value": 0.5,
     "unit": "m",
     "source": "Input"
@@ -839,9 +845,9 @@
     "unit": "\u00b0",
     "source": "Input"
   },
-  "div_open": {
-    "name": "Divertor open/closed configuration",
-    "value": false,
+  "div_baffle_type": {
+    "name": "Divertor baffle configuration (arc shaped, leg flux shaped, open)",
+    "value": "circle_baffle",
     "unit": "dimensionless",
     "source": "Input"
   },
